Suppress a ERROR message at PostgreSQL server shutdown.
Currently, collector (wait event) samples very high frequency.
So when Postgres received shutdown request, after the sampling fails (because of
force shutdown), the connection attempt to connect and sample again before checking
the shutdown receipt flag will result in an error.

In order to prevent this from happening, I added a process to check if the connection
is possible by performing PQping() when the result of PQstatus() is CONNECTION_BAD.
Since this check is not performed during normal operation, it is assumed that the
overhead will be light.

Note that this check might be useful for do_connect (for normal collector), but
it's only applied to the wait event collector.