Do not check log_min_messages
When Postgres is started with the -C option for variables marked as
GUC_RUNTIME_COMPUTED, it internally sets log_min_messages to
FATAL. This setting causes pg_statsinfo to exit server as the setting
being undesirable. Given this isn't a critical issue and a feasible
countermeasure is not found, we've decided to cease checking the
variable's value. Relevant instructions reflecting this change have
been added to the Japanese documentation.

(internal: ff464f2, fix #127)