Remove version-detection code
This commit removes the server version detection code present in the
function statsrepo.get_query_activity_statements. This part of the
code triggers errors when 'standard_conforming_strings' was set to
'off' and 'escape_string_warning' was 'on' due to the presence of
unescaped backslashes in string literals. With pg_statsinfo no longer
supporting multiple PostgreSQL versions with a single binary, this
version-detection code is deemed unnecessary and has been removed.

(internal: 9939556, fix #115)