Move location of pid file
The current location of the pid file may lead to the wrong processes
being killed by mistake due to the pid file being left behind when the
OS is restarted. This occurs especially when PostgreSQL is started
with the published unit file, which forcefully kills pg_statsinfo when
the service is stopped. To prevent this, move the pid file to the
location that is cleaned up at reboot.