Avoid division by zero errors
In some expressions used in SQL statements, some data anomaly—such as
counter resets—can occasionally result in error due to dividing zero
by zero in percentage calculations. In these SQL statements, such
errors are harmful, and it is more practical to avoid them by
returning a fallback value instead. This commit introduces a new
division function that returns 0 when both the divisor and dividend
are zero. It will still error out if only the divisor is zero.