Avoid division by zero errors
pg_statsinfo uses percentage calculations. Although zero divisors are
not expected in these calculations, cases have been found where both
the divisor and the dividend are zero under certain conditions, such
as after counter resets. In these cases, it's more beneficial to treat
the result as zero rather than erroring out. However, it will still be
treated as an error when the divisor is zero or negative and the
dividend is non-zero.

(internal: 40be32d, issue #159)