Chnage version number to 14.3 to 14.4
This commit corrects the return value of statsrepo.get_version() from
140300 to 140304. Recent changes related to minor updates were made
based on a misunderstanding that only the first two digits represent
the schema version, whereas in fact, the schema version is represented
by the first four digits. As a result, these changes unnecessarily led
to users needing to recreate the repository schema after minor updates
of pg_statsinfo.

The difference in the minor version number of pg_statsinfo does not
affect schema compatibility. In this commit, the schema version part
remains unchanged, and only the minor version part is updated to the
correct value.

(internal: a5d5a8f, issue #172)