Chnage version number to 15.2 to 15.3
This commit corrects the return value of statsrepo.get_version() from
150200 to 150203. Recent changes related to minor updates were made
based on a misunderstanding that only the first two digits represent
the schema version, whereas in fact, the schema version is represented
by the first four digits. As a result, these changes unnecessarily led
to users needing to recreate the repository schema after minor updates
of pg_statsinfo.

The difference in the minor version number of pg_statsinfo does not
affect schema compatibility. In this commit, the schema version part
remains unchanged, and only the minor version part is updated to the
correct value.

(internal: c95fe36, issue #172)