Remove queryids_lock
A normal backend should only write in a dedicated array position at the very
beginning of the query execution, and the underlying parallel workers will only
read from the same array position during the end of query execution.  As a
consequence there's no need to protect array access with a lock.

Note that this locking approach can lead to severe performance degradation in
some light OLTP workload with a high number of clients.